First commit.
This adds in support for 'get' and 'set' operations.

The unit tests don't yet cover everything. I need to extend them to the web API.

The implementation does not handle failure on the HTTP requests. It should.

The config.py is left out, which specifies 'host', 'path', and 'db'. Those are just strings, but I don't want to encode paths to our server in my code on Github.

This serves as a starting point. It is probably enough to get students started, though.